+++ util/coding.h
@@ -186,7 +186,7 @@ inline void PutLengthPrefixedSliceParts(std::string* dst,
                                         const SliceParts& slice_parts) {
   uint32_t total_bytes = 0;
   for (int i = 0; i < slice_parts.num_parts; ++i) {
-    total_bytes += slice_parts.parts[i].size();
+    total_bytes += static_cast<uint32_t>(slice_parts.parts[i].size());
   }
   PutVarint32(dst, total_bytes);
   for (int i = 0; i < slice_parts.num_parts; ++i) {
